requirements.txt
fastapi
passlib
uvicorn
dotenv
psycopg
SQLAlchemy
PyJWT